Remove email notification (it was called subscriptions) stuff
Not sure if restoring this at all makes sense. It was broken before in
an annoying way - it always sent a user one mail for each of her
subscriptions even when only one of them had changed. Even though the
one-click unsubscribe functionality was done nicely, it might make
more sense to look into rss to monitor changes to pages.